requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
rich>=13.7.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
